circular import nightmare; most tests are working again
