+++ mbuild/__init__.py
@@ -1,10 +1,7 @@
-from mbuild.atom import Atom
-from mbuild.bond import Bond
 from mbuild.box import Box
 from mbuild.coordinate_transform import *
-from mbuild.compound import *
-from mbuild.port import Port
+from mbuild.core import *
+from mbuild.topology import Topology
 
-from mbuild.formats import *
 from mbuild.tools import *
