+++ mbuild/examples/solvate/solvate.py
@@ -14,17 +14,12 @@ def main():
     print("Host (ethane) box: {}".format(host_box))
 
     # Create a water box.
-    water = mb.Compound()
-    water.append_from_file(get_fn('spc216.pdb'))
+    water = mb.load(get_fn('spc216.pdb'))
     guest_box = water.boundingbox()
     print("Guest (water) box: {}".format(guest_box))
 
     # Solvate ethane with water box.
     mb.solvate(ethane, water, host_box, guest_box)
-
-    ethane = ethane.to_trajectory()
-    ethane.topology.load_ff_bonds()
-    ethane.save(filename='ethane.hoomdxml')
 
 if __name__ == "__main__":
     main()
