+++ mbuild/tests/test_compound.py
@@ -1,7 +1,7 @@
 import pytest
 
 import mbuild as mb
-from mbuild.components.small_groups.ch3 import Ch3
+from mbuild.components.small_groups.ch3 import CH3
 from mbuild.testing.tools import get_fn
 from mbuild.tests.base_test import BaseTest
 
@@ -12,7 +12,7 @@ class TestCompound(BaseTest):
         mb.load(get_fn('methyl.pdb'))
 
     def test_update_from_file(self):
-        methyl = Ch3()
+        methyl = CH3()
         methyl.update_coordinates(get_fn("methyl.pdb"))
 
     def test_save(self):
@@ -44,14 +44,14 @@ class TestCompound(BaseTest):
         all_bonds = ethane.bond_list_by_kind()
         assert len(all_bonds) == 7
 
-    def test_atoms_in_range(self, ethane):
-        group = ethane.atoms_in_range(ethane.atoms[0].pos, 0.141)
-        assert sum([1 for x in group if x.kind == 'H']) == 3
-        assert sum([1 for x in group if x.kind == 'C']) == 2
-
-        group = ethane.atoms_in_range(ethane.atoms[0].pos, 0.141, max_items=4)
-        assert sum([1 for x in group if x.kind == 'H']) == 3
-        assert sum([1 for x in group if x.kind == 'C']) == 1
+    # def test_atoms_in_range(self, ethane):
+    #     group = ethane.atoms_in_range(ethane.atoms[0].pos, 0.141)
+    #     assert sum([1 for x in group if x.kind == 'H']) == 3
+    #     assert sum([1 for x in group if x.kind == 'C']) == 2
+    #
+    #     group = ethane.atoms_in_range(ethane.atoms[0].pos, 0.141, max_items=4)
+    #     assert sum([1 for x in group if x.kind == 'H']) == 3
+    #     assert sum([1 for x in group if x.kind == 'C']) == 1
 
     def test_remove(self, ethane):
         hydrogens = ethane.atom_list_by_kind("H")
