--- mbuild/tests/test_mbuild.py
@@ -1,13 +1,0 @@
-from mbuild.tests.base_test import BaseTest
-
-class TestMbuild(BaseTest):
-
-
-    def setUp(self):
-        pass
-
-    def test_something(self):
-        pass
-
-    def tearDown(self):
-        pass
