+++ mbuild/tools/__init__.py
@@ -1,4 +1,11 @@
-from mbuild.tools.tiled_compound import TiledCompound
+__all__ = ['apply_mask', 'random_mask_2d', 'random_mask_3d', 'sphere_mask',
+           'grid_mask_2d', 'grid_mask_3d', 'disk_mask',
+
+           'TiledCompound', 'Polymer',
+
+           'solvent_box', 'solvate']
+
 from mbuild.tools.mask import *
-from mbuild.tools.solvent import *
-from mbuild.tools.polymer import *
+from mbuild.tools.solvent import solvent_box, solvate
+from mbuild.tools.tiled_compound import TiledCompound
+from mbuild.tools.polymer import Polymer
