+++ mbuild/tools/parameterize/__init__.py
@@ -1,1 +1,2 @@
 from mbuild.tools.parameterize.atomtyper import find_atomtypes
+from mbuild.tools.parameterize.forcefield import apply_forcefield
