+++ mbuild/tools/validation.py
@@ -1,4 +1,3 @@
-
 def assert_port_exists(port_name, compound):
     """Ensure that a Port label exists in a Compound.  """
     if port_name in compound.labels:
